+++ xoa/accessors.py
@@ -20,6 +20,7 @@ xarray and pandas xoa accessors
 
 import warnings
 
+from . import exceptions
 from .misc import ERRORS
 
 
@@ -743,3 +744,58 @@ def register_xoa_accessors(name='xoa'):
         dataarrays={name: XoaDataArrayAccessor},
         datasets={name: XoaDatasetAccessor},
     )
+
+
+def register_accessors(xoa=True, xcf=False, xmeta=False, decode_sigma=False):
+    """Register xarray accessors
+
+    Parameters
+    ----------
+    xoa: bool, str
+        Register the main accessors with
+        :func:`~xoa.meta.register_xoa_accessors`.
+    xcf: bool, str
+        Register the :mod:`xoa.meta` module accessors with
+        :func:`~xoa.meta.register_meta_accessors` using the deprecated name "xcf".
+
+        .. deprecated::
+            Use ``xmeta`` instead.
+    xmeta: bool, str
+        Register the :mod:`xoa.meta` module accessors with
+        :func:`~xoa.meta.register_meta_accessors`.
+    decode_sigma: bool, str
+        Register the :mod:`xoa.sigma` module accessor with
+        :func:`~xoa.meta.register_sigma_accessor`.
+
+    See also
+    --------
+    xoa.accessors
+    """
+    if xoa:
+        from .accessors import register_xoa_accessors
+
+        kw = {"name": xoa} if isinstance(xoa, str) else {}
+        register_xoa_accessors(**kw)
+    if xcf:
+        from .accessors import register_meta_accessors
+
+        exceptions.woa_warn(
+            "The 'xcf' parameter is deprecated. Use 'xmeta' instead.",
+            "deprecation",
+            stacklevel=2,
+        )
+        kw = {"name": xcf} if isinstance(xcf, str) else {}
+        # Register with "xcf" name for backward compatibility
+        if not kw:
+            kw = {"name": "xcf"}
+        register_meta_accessors(**kw)
+    if xmeta:
+        from .accessors import register_meta_accessors
+
+        kw = {"name": xmeta} if isinstance(xmeta, str) else {}
+        register_meta_accessors(**kw)
+    if decode_sigma:
+        from .accessors import register_sigma_accessor
+
+        kw = {"name": decode_sigma} if isinstance(decode_sigma, str) else {}
+        register_sigma_accessor(**kw)
